fix handling of tie-isomorphic states
The bug that 9d2de34e61 almost fixed was that the move-searching
algorithm treats isomprohic board states the same (e.g., mirror
image), but that's not right for tie states (which are the previous
states of the game). In particular, when a move starts in a state
isomorphic to a previous state, the move search shouldn't immediately
give up on the grounds that a tie is reached already.